Add chrono timing
Not perfect, just a quick addition that includes the ETWUtils::start and ETWUtils::stop calls in the time. Will improve later

Some internal changes, increased the sleep after spinning up the ProcessTrace thread.

TODO: wait on a locked mutex which unlocks once the callback function receives the CollectionStart event, move timing to internals of ETWUtils to be done in the callback at CollectionStart and CollectionEnd