Better timing & Sleep fix
Fixed hacky Sleep workaround to benchmark starting before trace callback was running, uses an internal mutex to wait for the session to start tracing before allowing the benchmarking thread to continue execution.

Timing sessions was moved to begin when the tracing begins, in the trace callback.

Overall timings should be slightly better.